Implementation of XOR Gate using Backpropagation in Neural Networks
Error Back Propagation MLP
The Error Backward Propagation or backpropagation is a common method of training artificial
neural networks and used in conjunction with an optimization method such as gradient descent.
The algorithm repeats a two phase cycle, propagation and weight update. When an input vector
is presented to the network, it is propagated forward through the network, layer by layer, until it
reaches the output layer. The output of the network is then compared to the desired output, using
a loss function, and an error value is calculated for each of the neurons in the output layer. The
error values are then propagated backwards, starting from the output, until each neuron has an
associated error value which roughly represents its contribution to the original output.
Backpropagation uses these error values to calculate the gradient of the loss function with respect
to the weights in the network. In the second phase, this gradient is fed to the optimization
method, which in turn uses it to update the weights, in an attempt to minimize the loss function.
The importance of this process is that, as the network is trained, the neurons in the intermediate
layers organize themselves in such a way that the different neurons learn to recognize different
characteristics of the total input space. After training, when an arbitrary input pattern is present
which contains noise or is incomplete, neurons in the hidden layer of the network will respond
with an active output if the new input contains a pattern that resembles a feature that the
individual neurons have learned to recognize during their training.
Backpropagation requires a known, desired output for each input value in order to calculate the
loss function gradient – it is therefore usually considered to be a supervised learning method;
nonetheless, it is also used in some unsupervised networks such as auto encoders. It is a
generalization of the delta rule to multi-layered feedforward networks, made possible by using the chain rule to iteratively compute gradients for each layer. Backpropagation requires that the activation function used by the artificial neurons (or "nodes") be differentiable.