To design and implement a Mobile price range classifier.
Theory:
To design a neural network following steps are followed.

1. Preprocess and load data- Data is the key for the working of neural network and we need to
process it before feeding to the neural network. In this step, we will also visualize data which
will help us to gain insight into the data.
2. Define model- Now we need a neural network model. This means we need to specify the
number of hidden layers in the neural network and their size, the input and output size.
3. Loss and optimizer- Now we need to define the loss function according to our task. We also
need to specify the optimizer to use with learning rate and other hyperparameters of the
optimizer.
4. Fit model- This is the training step of the neural network. Here we need to define the number
of epochs for which we need to train the neural network.

After fitting the model, we can test it on test data to check whether the case of overfitting. We
can save the weights of the model and use it later whenever required.