To design and implement a CNN for Multiclass Object classification on image data.
Convolutional Neural Networks (CNNs / ConvNets)
Convolutional neural networks as very similar to the ordinary feed-forward neural networks.
They differ in the sense that CNNs assume explicitly that the inputs are images, which enables us
to encode specific properties in the architecture to recognize certain patterns in the images. The
CNNs make use of spatial nature of the data. It means, CNNs perceive the objects similar to our
perception of different objects in nature. For example, we recognize various objects by their
shapes, size and colors. These objects are combinations of edges, corners, color patches, etc.
CNNs can use a variety of detectors (such as edge detectors, corner detectors) to interpret
images. These detectors are called filters or kernels. The mathematical operator that takes an
image and a filter as input and produces a filtered output (e.g. edges, corners, etc. ) is
called convolution.