To implement concept of transfer learning
Transfer Learning is a machine learning method where we reuse a pre-trained model as the
starting point for a model on a new task.A model trained on one task is repurposed on a second,
related task as an optimization that allows rapid progress when modeling the second task.By
applying transfer learning to a new task, one can achieve significantly higher performance than
training with only a small amount of data. Transfer learning is so common that it is rare to train a
model for an image or natural language processing-related tasks from scratch.Instead,
researchers and data scientists prefer to start from a pre-trained model that already knows how
to classify objects and has learned general features like edges, shapes in images.ImageNet,
AlexNet, and Inception are typical examples of models that have the basis of Transfer learning.