To implement Generative Adversarial network
What Are Generative Adversarial Network?
Generative Adversarial Networks or GANs are a deep-learning-based generative model that is
used for Unsupervised Learning. It is basically a system where two competing Neural
Networks compete with each other to create or generate variations in the data.
It was first described in a paper in 2014 by Ian Goodfellow and a standardized and much stable
model theory was proposed by Alec Radford in 2016 which is known as DCGAN (Deep
Convolutional General Adversarial Networks). Most of the GANs today in existence use
DCGAN architecture.
The GANs architecture consists of two sub-models known as the Generator model and
the Discriminator Model. Let us try to understand how GANs actually work.